% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_boxplot_star.r
\name{geom_boxplot_star}
\alias{geom_boxplot_star}
\title{geom_boxplot significance}
\usage{
geom_boxplot_star(stat.tests, ranges, rm.outliers = TRUE,
  categories = cancerTypes)
}
\arguments{
\item{stat.tests}{}

\item{ranges}{}

\item{rm.outliers}{}

\item{categories}{}
}
\value{
geom_boxplot_star
}
\description{
Creates the ggplot layer to display a significance measure on a geom_boxplot.
}

