% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_bar.R
\name{star_bar}
\alias{star_bar}
\title{geom_bar significance}
\usage{
star_bar(gg, significance, fit = TRUE, dist = 0)
}
\arguments{
\item{gg}{ggplot with a geom_bar to add significance marks.}

\item{significance}{logical vector with the significant pairs.}

\item{fit}{logical value indicating if the arc should fit the shape of the data.}

\item{dist}{numerical value indicating the distance from the arc to the data.}
}
\value{
gg ggplot with the geom_bar and the significance marks.
}
\description{
Adds significance marks to a ggplot with a geom_bar.
}

