% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_box.R
\name{star_box}
\alias{star_box}
\title{geom_boxplot significance}
\usage{
star_box(gg, significance, rm.outliers = TRUE, fit = FALSE, dist = 0.05)
}
\arguments{
\item{gg}{ggplot with a geom_boxplot to add significance marks.}

\item{significance}{logical vector with the significant pairs.}

\item{rm.outliers}{logical value the arc should fit the data with or without the outliers.}

\item{fit}{logical value indicating if the arc should fit the shape of the data.}

\item{dist}{numerical value indicating the distance from the arc to the data.}
}
\value{
gg ggplot with the geom_boxplot and the significance marks.
}
\description{
Adds significance marks to a ggplot with a geom_boxplot.
}

