% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/star_boxplot.R
\name{star_boxplot}
\alias{star_boxplot}
\title{geom_boxplot significance}
\usage{
star_boxplot(gg, significance)
}
\arguments{
\item{gg}{ggplot with a geom_boxplot to add significance marks.}

\item{significance}{logical vector with the significant pairs.}
}
\value{
gg ggplot with the geom_boxplot and the significance marks.
}
\description{
Adds significance marks to a ggplot with a geom_boxplot.
}

