% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/star_bar.R
\name{star_bar}
\alias{star_bar}
\title{star_bar}
\usage{
star_bar(gg, significance)
}
\arguments{
\item{gg}{ggplot with a geom_bar to add significance marks.}

\item{significance}{logical vector with the significant pairs.}
}
\value{
gg ggplot with the geom_bar and the significance marks.
}
\description{
Adds significance marks to a ggplot with a geom_bar.
}

